Issue-80 - refactored Media Buttons to to extract common data into super class
