+++ app/src/main/java/com/example/mike/mp3player/client/views/buttons/SkipToNextButton.java
@@ -1,4 +1,4 @@
-package com.example.mike.mp3player.client.views;
+package com.example.mike.mp3player.client.views.buttons;
 
 import android.content.Context;
 import android.graphics.drawable.Drawable;
@@ -12,25 +12,17 @@ import com.example.mike.mp3player.client.MediaControllerAdapter;
 import javax.inject.Inject;
 import javax.inject.Named;
 
-public class SkipToNextButton {
-
-    private final Context context;
-    private final MediaControllerAdapter mediaControllerAdapter;
-    private final Handler mainUpdater;
-    private ImageView view;
+public class SkipToNextButton extends MediaButton{
 
     @Inject
     public SkipToNextButton(Context context, MediaControllerAdapter mediaControllerAdapter, @Named("main") Handler mainUpdater) {
-        this.context = context;
-        this.mediaControllerAdapter = mediaControllerAdapter;
-        this.mainUpdater = mainUpdater;
+        super(context, mediaControllerAdapter, mainUpdater);
     }
 
     public void init(ImageView view) {
         this.view = view;
         Drawable drawable = context.getDrawable(R.drawable.ic_baseline_skip_next_24px);
         this.view.setImageDrawable(drawable);
-//        this.view.init(R.drawable.ic_baseline_skip_previous_24px, 2);
         this.view.setOnClickListener(this::skipToNext);
     }
