+++ app/src/main/java/com/example/mike/mp3player/client/views/fragments/PlaybackToolbarExtendedFragment.java
@@ -12,9 +12,9 @@ import androidx.fragment.app.FragmentActivity;
 
 import com.example.mike.mp3player.R;
 import com.example.mike.mp3player.client.activities.MediaActivityCompat;
-import com.example.mike.mp3player.client.views.PlayPauseButton;
-import com.example.mike.mp3player.client.views.SkipToNextButton;
-import com.example.mike.mp3player.client.views.SkipToPreviousButton;
+import com.example.mike.mp3player.client.views.buttons.PlayPauseButton;
+import com.example.mike.mp3player.client.views.buttons.SkipToNextButton;
+import com.example.mike.mp3player.client.views.buttons.SkipToPreviousButton;
 
 
 @Deprecated
