+++ app/src/main/res/layout/fragment_test_seek_bar.xml
@@ -8,7 +8,7 @@
         android:layout_width="match_parent"
         android:layout_height="100dp" />
 
-    <com.example.mike.mp3player.client.views.PlayPauseButton
+    <com.example.mike.mp3player.client.views.buttons.PlayPauseButton
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:id="@+id/ppButton"/>
